+++ slash_commands/index-fttc.js
@@ -5,8 +5,7 @@ const TowerParser = require('../parser/tower-parser');
 
 const Parsed = require('../parser/parsed');
 
-const gHelper = require('../helpers/general.js');
-const Index = require('../helpers/index.js');
+const { displayOneOrMultiplePagesNew, genCompletionLink } = require('../helpers/index.js');
 
 const { paleorange } = require('../jsons/colors.json');
 
@@ -21,15 +20,10 @@ const tower2Option = new SlashCommandStringOption().setName('tower2').setDescrip
 const numTowerTypesOption = new SlashCommandIntegerOption()
     .setName('num_tower_types')
     .setDescription('Number of tower types')
+    .setMinValue(1)
     .setRequired(false);
 
 const personOption = new SlashCommandStringOption().setName('person').setDescription('Completer').setRequired(false);
-
-const reloadOption = new SlashCommandStringOption()
-    .setName('reload')
-    .setDescription('Do you need to reload completions from the index but for a much slower runtime?')
-    .setRequired(false)
-    .addChoices({ name: 'Yes', value: 'yes' });
 
 builder = new SlashCommandBuilder()
     .setName('fttc')
@@ -39,9 +33,10 @@ builder = new SlashCommandBuilder()
     .addStringOption(tower2Option)
     .addIntegerOption(numTowerTypesOption)
     .addStringOption(personOption)
-    .addStringOption(reloadOption);
 
 async function execute(interaction) {
+    await interaction.deferReply();
+
     const validationFailure = validateInput(interaction);
     if (validationFailure) {
         return await interaction.reply({
@@ -55,23 +50,103 @@ async function execute(interaction) {
         new Parsed()
     );
 
-    await interaction.deferReply();
+    return await displayFttcFilterAll(interaction, parsed)
+}
 
-    const forceReload = interaction.options.getString('reload') ? true : false;
-
-    const allCombos = await Index.fetchInfo('fttc', forceReload);
-
-    const mtime = Index.getLastCacheModified('fttc');
-
-    let filteredCombos = filterResults(allCombos, parsed);
-
-    if (filteredCombos.length == 0) {
-        const noCombosEmbed = new Discord.EmbedBuilder().setTitle(titleNoCombos(parsed)).setColor(paleorange);
-
-        return interaction.editReply({ embeds: [noCombosEmbed] });
-    } else {
-        return await embedOneOrMultiplePages(interaction, parsed, filteredCombos, mtime);
+async function displayFttcFilterAll(interaction, parsed) {
+    const fetchParams = getUrlParams(parsed)
+    if ((await fetchFttc(fetchParams)).count <= 0) {
+        return interaction.reply({
+            embeds: [
+                new Discord.EmbedBuilder().setTitle(titleNoCombos(parsed))
+            ]
+        })
     }
+    fetchParams.set('count', '10');
+    return await displayOneOrMultiplePagesNew(
+        interaction, fetchFttc, fetchParams,
+        includedColumns(parsed),
+        completion => {
+            const boldOg = (str) => completion.og ? `**${str}**` : str;
+
+            const towers = JSON.parse(completion.towerset).map(tower => {
+                const towerCanonical = Aliases.getCanonicalForm(tower);
+                const towerAbbreviation = FTTC_TOWER_ABBREVIATIONS[towerCanonical].toUpperCase();
+                return parsed.towers && parsed.towers.includes(towerCanonical)
+                    ? `**${towerAbbreviation}**`
+                    : towerAbbreviation;
+            }).join(' | ')
+
+            return {
+                'Towers': boldOg(towers),
+                'Map': boldOg(completion.map),
+                'Person': boldOg(completion.person),
+                'Link': boldOg(genCompletionLink(completion))
+            };
+        },
+        embed => {
+            embed
+            .setTitle(title(parsed))
+            .setColor(paleorange)
+            .setFooter({ text: `---\nAny OG completion(s) bolded` });
+        }
+    );
+}
+
+function includedColumns(parsed) {
+    // Setup / Data consolidation
+    let displayCols = ['Towers', 'Map', 'Person', 'Link'];
+
+    if (parsed.person) {
+        displayCols = displayCols.filter((col) => col != 'Person');
+    }
+
+    if (parsed.map) {
+        displayCols = displayCols.filter((col) => col != 'Map');
+    }
+
+    if (displayCols.length === 4) {
+        displayCols = displayCols.filter((col) => col != 'Person');
+    }
+
+    return displayCols;
+}
+
+async function fetchFttc(searchParams) {
+    let res = await fetch('https://btd6index.win/fetch-fttc?' + searchParams);
+    let resJson = await res.json();
+    if ('error' in resJson) {
+        throw new Error(resJson.error);
+    }
+    return resJson;
+}
+
+function getUrlParams(parsed) {
+    params = new URLSearchParams()
+
+    if (parsed.map) {
+        params.set('map', Aliases.toIndexNormalForm(parsed.map))
+    }
+
+    if (parsed.natural_number) {
+        params.set('towercount', parsed.natural_number)
+    }
+
+    if (parsed.towers) {
+        params.set('towerincludes', parsed.towers.map(t => `"${t}"`))
+    }
+
+    if (parsed.person) {
+        params.set('person', parsed.person)
+    }
+
+    if(keepOnlyOG(parsed)) {
+        params.set('og', 1)
+    }
+
+    params.set('count', 0)
+
+    return params
 }
 
 ////////////////////////////////////////////////////////////
@@ -142,11 +217,7 @@ function validateInput(interaction) {
     }
 
     if (parsedTower2.hasErrors()) {
-        return 'Tower2 did not match a towe';
-    }
-
-    if (parsedNumTowerTypes.hasErrors()) {
-        return `Number of Combos must be >= 1`;
+        return 'Tower2 did not match a tower';
     }
 
     const parsedTowers = parsedTower1.merge(parsedTower2);
@@ -162,36 +233,8 @@ function validateInput(interaction) {
     }
 }
 
-function filterResults(allCombos, parsed) {
-    results = allCombos;
-
-    if (parsed.map) {
-        results = results.filter((combo) => Aliases.toAliasNormalForm(combo.MAP) == parsed.map);
-    }
-
-    if (parsed.natural_number) {
-        results = results.filter((combo) => combo.TOWERS.length === parsed.natural_number);
-    }
-
-    if (parsed.person) {
-        results = results.filter((combo) => {
-            return combo.PERSON.toLowerCase().split(' ').join('_') === parsed.person.toLowerCase().split(' ').join('_');
-        });
-    }
-
-    if (parsed.towers) {
-        results = results.filter((combo) => parsed.towers.every((specifiedTower) => combo.TOWERS.includes(specifiedTower)));
-    }
-
-    if (keepOnlyOG(parsed) || !parsed.hasAny()) {
-        results = results.filter((combo) => combo.OG);
-    }
-
-    return results;
-}
-
 function keepOnlyOG(parsed) {
-    return parsed.natural_number && !parsed.person && !parsed.tower;
+    return !parsed.hasAny() || (parsed.natural_number && !parsed.person && !parsed.tower);
 }
 
 ////////////////////////////////////////////////////////////
@@ -223,89 +266,19 @@ const FTTC_TOWER_ABBREVIATIONS = {
     beast_handler: 'bst'
 };
 
-async function embedOneOrMultiplePages(interaction, parsed, combos, mtime) {
-    // Setup / Data consolidation
-    let displayCols = ['TOWERS', 'MAP', 'PERSON', 'LINK'];
-
-    if (parsed.person) {
-        displayCols = displayCols.filter((col) => col != 'PERSON');
-    }
-
-    if (parsed.map) {
-        displayCols = displayCols.filter((col) => col != 'MAP');
-    }
-
-    if (displayCols.length === 4) {
-        displayCols = displayCols.filter((col) => col != 'PERSON');
-    }
-    const colData = Object.fromEntries(
-        displayCols.map((col) => {
-            if (col == 'TOWERS') {
-                const boldedAbbreviatedTowers = combos.map((combo) =>
-                    combo[col].map((tower) => {
-                        if (tower) {
-                            const towerCanonical = Aliases.getCanonicalForm(tower);
-                            const towerAbbreviation = FTTC_TOWER_ABBREVIATIONS[towerCanonical].toUpperCase();
-                            return parsed.towers && parsed.towers.includes(towerCanonical)
-                                ? `**${towerAbbreviation}**`
-                                : towerAbbreviation;
-                        }
-                    })
-                );
-                const colValues = boldedAbbreviatedTowers.map((comboTowers, i) => {
-                    let value = comboTowers.join(' | ');
-                    if (combos[i].OG && !parsed.towers) {
-                        value = `**${value}**`;
-                    }
-                    return value;
-                });
-                return [col, colValues];
-            } else {
-                const colValues = combos.map((combo) => {
-                    value = combo[col];
-                    if (combo.OG) {
-                        value = `**${value}**`;
-                    }
-                    return value;
-                });
-                return [col, colValues];
-            }
-        })
-    );
-
-    const numOGCompletions = combos.filter((combo) => combo.OG).length;
-
-    function setOtherDisplayFields(challengeEmbed) {
-        challengeEmbed
-            .setTitle(embedTitle(parsed, combos))
-            .setColor(paleorange)
-            .setDescription(`Index last reloaded ${gHelper.timeSince(mtime)} ago`);
-
-        if (numOGCompletions == 1) {
-            challengeEmbed.setFooter({ text: `---\nOG completion bolded` });
-        }
-        if (numOGCompletions > 1) {
-            challengeEmbed.setFooter({
-                text: `---\n${numOGCompletions} OG completions bolded`
-            });
-        }
-    }
-
-    return await Index.displayOneOrMultiplePages(interaction, colData, setOtherDisplayFields);
-}
-
-function embedTitle(parsed, combos) {
-    t = combos.length > 1 ? 'All FTTC Combos ' : 'Only FTTC Combo ';
-    if (parsed.person) t += `by ${combos[0].PERSON} `;
+function title(parsed) {
+    // t = combos.length > 1 ? 'All FTTC Combos ' : 'Only FTTC Combo ';
+    let t = 'All FTTC Combos '
+    if (parsed.person) t += `by ${parsed.person} `;
     if (parsed.natural_number) t += `with ${parsed.natural_number} towers `;
-    if (parsed.map) t += `on ${combos[0].MAP} `;
+    if (parsed.map) t += `on ${Aliases.toIndexNormalForm(parsed.map)} `;
     if (parsed.towers) t += `including ${Towers.towerUpgradeToIndexNormalForm(parsed.towers[0])} `;
     if (parsed.towers && parsed.towers[1]) t += `and ${Towers.towerUpgradeToIndexNormalForm(parsed.towers[1])} `;
     return t.slice(0, t.length - 1);
 }
 
 function titleNoCombos(parsed) {
-    t = 'No FTTC Combos Found ';
+    let t = 'No FTTC Combos Found ';
     if (parsed.person) t += `by "${parsed.person}" `;
     if (parsed.natural_number) t += `with ${parsed.natural_number} towers `;
     if (parsed.map) t += `on ${Aliases.toIndexNormalForm(parsed.map)} `;
